gentrace-spike: elaborates Spike traces
Elaborates Spike traces with all the information TBM needs. The disassembler
part is a bit of a hack. It should be possible to extend Spike to output the
registers each instruction reads, instead of just the registers they write (as
it does for memory accesses) and then the disassembler will not be required.

Change-Id: I427d87aad7725903c93b5b0fab8b0bb22866b458
GitOrigin-RevId: cae643d26e7f1a2201038b09ac8d761ddd0ccd6e